% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airfoil.R
\docType{data}
\name{airfoil}
\alias{airfoil}
\title{airfoil self noise dataset}
\format{An object of class \code{data.frame} with 1503 rows and 6 columns.}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/Airfoil+Self-Noise#}{UCI Archive}
}
\usage{
data(airfoil)
}
\description{
The NASA data set comprises different size NACA 0012 airfoils at various wind tunnel speeds and angles of attack.
The span of the airfoil and the observer position were the same in all of the experiments.
}
\examples{
data(airfoil)
airfoil$pressure
}
\references{
T.F. Brooks, D.S. Pope, and A.M. Marcolini.
Airfoil self-noise and prediction.
Technical report, NASA RP-1218, July 1989.

K. Lau.
A neural networks approach for aerofoil noise prediction.
Master's thesis, Department of Aeronautics.
Imperial College of Science, Technology and Medicine (London, United Kingdom), 2006.

Lopez, R. and Balsa-Canto, E. and Onate, E.
Neural Networks for Variational Problems in Engineering.
PhD Thesis, Technical University of Catalonia, 2008.
}
\keyword{datasets}
