% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ghosal, Hooker 2018.R
\name{boostStage}
\alias{boostStage}
\title{generate stage 2 (and more) RF for Ghosal, Hooker RF implementation}
\usage{
boostStage(
  rf,
  formula = NULL,
  train_data = NULL,
  pred_data = NULL,
  num_trees = num_trees,
  min_node_size = NULL,
  m_try = NULL,
  keep_inbag = TRUE,
  intervals = TRUE,
  alpha = alpha,
  weights = NULL,
  num_stages = 2,
  prop = prop,
  num_threads = num_threads
)
}
\description{
This function is primarily meant to be used within the GhosalBoostRF() function. All parameters are same as in GhosalBoostRF().
}
\examples{
boostStage <- function(rf, formula = NULL, train_data = NULL, pred_data = NULL, num_trees = num_trees, 
min_node_size = NULL, m_try = NULL, keep_inbag = TRUE, 
intervals = TRUE, 
alpha = alpha, forest_type = forest_type, weights = NULL, num_stages = 2, 
replace = replace, prop = prop, quantreg = TRUE, num_threads = num_threads)
}
\keyword{cats}
