% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roy, Larocque 2019.R
\name{genCHDInt}
\alias{genCHDInt}
\title{generates BOP contiguous HDI prediction intervals from Roy, Larocque 2019}
\usage{
genCHDInt(BOP, alpha = alpha)
}
\arguments{
\item{BOP}{BOP object generated from genBOP() function.}
}
\description{
This function is primarily meant to be used within the RoyRF() function.
}
\keyword{calibration}
\keyword{forest,}
\keyword{random}
