% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ghosal_Hooker_2018.R
\name{genCombRF}
\alias{genCombRF}
\title{Generates stage 1 RF for Ghosal, Hooker RF implementation. Helper function.}
\usage{
genCombRF(
  formula = NULL,
  train_data = NULL,
  pred_data = NULL,
  num_trees = num_trees,
  min_node_size = NULL,
  m_try = NULL,
  keep_inbag = TRUE,
  intervals = TRUE,
  alpha = NULL,
  importance = "none",
  weights = NULL,
  prop = NULL,
  inbag = NULL,
  num_threads = NULL
)
}
\description{
This function is primarily meant to be used within GhosalBoostRF().
}
\keyword{internal}
