% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roy_Larocque_2019.R
\name{genqInt}
\alias{genqInt}
\title{Generates BOP quantile prediction intervals from Roy, Larocque 2019.}
\usage{
genqInt(BOP, alpha = alpha, interval_type = interval_type)
}
\description{
This function is primarily meant to be used within the RoyRF() function.
}
\keyword{internal}
