% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tung, Huang, Nyugen, Khan 2014.R
\name{genRF}
\alias{genRF}
\title{generate quantile RF}
\usage{
genRF(
  formula = NULL,
  train_data = NULL,
  pred_data = NULL,
  num_trees = num_trees,
  min_node_size = NULL,
  m_try = NULL,
  keep_inbag = TRUE,
  intervals = TRUE,
  alpha = alpha,
  forest_type = "RF",
  importance = "none",
  weights = NULL,
  num_threads = num_threads
)
}
\description{
This function is primarily meant to be used within the TungUbRF() function. All parameters are same as in TungUbRf().
}
\keyword{forest}
\keyword{random}
