% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ghosal, Hooker 2018.R
\name{GHVar}
\alias{GHVar}
\title{generate prediction intervals for Ghosal, Hooker 2018 implementation.}
\usage{
GHVar(
  boostRF,
  train_data,
  pred_data,
  variant,
  dep,
  alpha,
  num_threads = num_threads
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function is primarily meant to be used within the GhosalBoostRF() function. All parameters are same as in GhosalBoostRF().
}
\examples{
GHVar <- function(boostRF, train_data, pred_data, variant, dep, alpha, num_threads = num_threads)
}
\keyword{U}
\keyword{forest,}
\keyword{intervals}
\keyword{prediction}
\keyword{random}
\keyword{statistics,}
