% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{parse.formula}
\alias{parse.formula}
\title{Authors: Marvin N. Wright, Gregor DeCillia}
\usage{
parse.formula(formula, data, env = parent.frame())
}
\arguments{
\item{formula}{Object of class \code{formula} or \code{character} describing the model to fit.}

\item{data}{Training data of class \code{data.frame}.}

\item{env}{The environment in which the left hand side of \code{formula} is evaluated.}
}
\value{
Dataset including selected columns and interactions.
}
\description{
taken from source code for ranger package; not exported with package
}
\details{
Parse formula and return dataset containing selected columns.
Interactions are supported for numerical columns only.
An interaction column is the product of all interacting columns.
}
