\name{PIRF-package}
\alias{PIRF-package}
\alias{PIRF}
\docType{package}
\title{
\packageTitle{PIRF}
}
\description{
\packageDescription{PIRF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PIRF}
\packageIndices{PIRF}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PIRF}

Maintainer: \packageMaintainer{PIRF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
