% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirf.R
\name{pirf}
\alias{pirf}
\title{pirf()}
\usage{
pirf(
  formula = NULL,
  train_data = NULL,
  method = "oob",
  alpha = 0.1,
  symmetry = TRUE,
  seed = NULL,
  m_try = NULL,
  num_trees = 500,
  min_node_size = NULL,
  num_threads = parallel::detectCores(),
  calibrate = FALSE,
  Roy_method = "quantile",
  featureBias = FALSE,
  predictionBias = TRUE,
  Tung_R = 5,
  Tung_num_trees = 75,
  variant = 1,
  Ghosal_num_stages = 2,
  prop = 0.618,
  concise = TRUE,
  interval_type = "two-sided",
  ...
)
}
\arguments{
\item{formula}{Object of class formula or character describing the model to fit. Interaction terms supported only for numerical variables.}

\item{train_data}{Training data of class data.frame.}

\item{method}{Choose what method to generate RF prediction intervals. Options are \code{method = c("oob", "quantile", "cqrf", "brf", "bop", "bcqrf", "hdi")}. Defaults to \code{method = "oob"}.}

\item{alpha}{Significance level for prediction intervals. Defaults to \code{alpha = 0.1}.}

\item{symmetry}{True if constructing symmetric out-of-bag prediction intervals, False otherwise. Used only \code{method = "oob"}. Defaults to \code{symmetry = TRUE}.}

\item{seed}{Seed for random number generation. Currently not utilized.}

\item{m_try}{Number of variables to randomly select from at each split.}

\item{num_trees}{Number of trees used in the random forest.}

\item{min_node_size}{Minimum number of observations before split at a node.}

\item{num_threads}{The number of threads to use in parallel. Default is the current number of cores.}

\item{calibrate}{If \code{calibrate = TRUE}, intervals are calibrated to achieve nominal coverage. Currently uses quantiles to calibrate. Only for \code{method = "bop"}.}

\item{Roy_method}{Interval method for \code{method = "bop"}.
Options are \code{Roy_method = c("quantile", "hdi", "CHDI")}.}

\item{featureBias}{Remove feature bias. Only for \code{method = "bcqrf"}.}

\item{predictionBias}{Remove prediction bias. Only for \code{method = "bcqrf"}.}

\item{Tung_R}{Number of repetitions used in bias removal. Only for \code{method = "bcqrf"}.}

\item{Tung_num_trees}{Number of trees used in bias removal. Only for \code{method = "bcqrf"}.}

\item{variant}{Choose which variant to use. Options are \code{method = c("1", "2")}. Only for \code{method = "brf"}.}

\item{Ghosal_num_stages}{Number of total stages. Only for \code{method = "brf"}.}

\item{prop}{Proportion of training data to sample for each tree. Only for \code{method = "brf"}.}

\item{concise}{If  concise = TRUE, only predictions output. Defaults to \code{concise = FALSE}.}

\item{interval_type}{Type of prediction interval to generate.
Options are \code{method = c("two-sided", "lower", "upper")}. Default is  \code{method = "two-sided"}.}

\item{test_data}{Test data of class data.frame. Utilizes ranger::predict() to produce prediction intervals for test data.}
}
\description{
Implements seven different random forest prediction interval methods.
}
\details{
The seven methods implemented are cited in the References section.
Additional information can be found within those references.
Each of these methods are implemented by utilizing the ranger package.
For \code{method = "oob"}, prediction intervals are generated using out-of-bag residuals.
\code{method = "cqrf"} utilizes a split-conformal approach.
\code{method = "bop"} uses a bag-of-predictors approach.
\code{method = "brf"} performs boosting to reduce bias in the random forest, and estimates variance.
The authors provide multiple variants to their methodology.
\code{method = "bcqrf"} debiases feature selection and prediction. Prediction intervals are generated using quantile regression forests.
\code{method = "hdi"} delivers prediction intervals through highest-density interval regression forests.
\code{method = "quantile"} utilizes quantile regression forests.
}
\seealso{
\link[ranger]{ranger}

\link[rfinterval]{rfinterval}

\link[piRF]{predict.pirf}
}
\author{
Chancellor Johnstone

Haozhe Zhang
}
