% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tung, Huang, Nyugen, Khan 2014.R
\name{predictionUbRF}
\alias{predictionUbRF}
\title{performs prediction debiasing from Tung 2014}
\usage{
predictionUbRF(
  rf,
  formula = NULL,
  train_data = NULL,
  pred_data = NULL,
  num_trees = NULL,
  min_node_size = NULL,
  m_try = NULL,
  keep_inbag = TRUE,
  intervals = TRUE,
  alpha = alpha,
  forest_type = "QRF",
  weights = NULL,
  num_threads = num_threads
)
}
\description{
This function is primarily meant to be used within the TungUbRF() function. All parameters are same as in TungUbRf().
}
\keyword{forest}
\keyword{random}
