% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ghosal_Hooker_2018.R
\name{boostStage}
\alias{boostStage}
\title{Generates stage 2 (and more) RF for Ghosal, Hooker RF implementation. Helper function.}
\usage{
boostStage(
  rf,
  formula = NULL,
  train_data = NULL,
  pred_data = NULL,
  num_trees = num_trees,
  min_node_size = NULL,
  m_try = NULL,
  keep_inbag = TRUE,
  intervals = TRUE,
  alpha = alpha,
  weights = NULL,
  num_stages = 2,
  prop = prop,
  num_threads = num_threads,
  variant = NULL
)
}
\description{
Used within GhosalBoostRF().
}
\keyword{internal}
