% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roy_Larocque_2019.R
\name{genBOP}
\alias{genBOP}
\title{Generate BOP sets from Roy, Larocque 2019}
\usage{
genBOP(
  rf,
  inbag = rf$inbag.counts,
  alpha = alpha,
  pred_data,
  train_data,
  num_threads = num_threads,
  calibrate = calibrate
)
}
\description{
This function is primarily meant to be used within the RoyRF() function. All parameters are same as in RoyRF().
}
\keyword{internal}
