% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roy_Larocque_2019.R
\name{genCHDInt}
\alias{genCHDInt}
\title{generates BOP contiguous HDI prediction intervals from Roy, Larocque 2019}
\usage{
genCHDInt(BOP, alpha = alpha)
}
\description{
This function is primarily meant to be used within the RoyRF() function.
}
\keyword{internal}
