% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tung_Huang_Nyugen_Khan_2014.R
\name{genWeights}
\alias{genWeights}
\title{Generate weights for RF through feature bias reduction method outlined in Tung 2014.}
\usage{
genWeights(
  formula = NULL,
  train_data = NULL,
  pred_data = NULL,
  feature_num_trees = feature_num_trees,
  min_node_size = NULL,
  m_try = NULL,
  keep_inbag = TRUE,
  intervals = TRUE,
  alpha = alpha,
  forest_type = "RF",
  importance = "permutation",
  R = R,
  num_threads = num_threads
)
}
\description{
This function is primarily meant to be used within the TungUbRF() function. All parameters are same as in TungUbRF().
}
\keyword{internal}
