% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ghosal_Hooker_2018.R
\name{GHVar}
\alias{GHVar}
\title{Generate prediction intervals for Ghosal, Hooker 2018 implementation. Helper function.}
\usage{
GHVar(
  boostRF,
  train_data,
  pred_data,
  variant,
  dep,
  alpha,
  num_threads = num_threads,
  interval_type = interval_type
)
}
\description{
This function is primarily meant to be used within GhosalBoostRF().
}
\keyword{internal}
