% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.R
\name{rfint}
\alias{rfint}
\title{rfint()}
\usage{
rfint(
  formula = formula,
  train_data = NULL,
  test_data = NULL,
  method = "Zhang",
  alpha = 0.1,
  symmetry = TRUE,
  seed = NULL,
  m_try = 2,
  num_trees = 500,
  min_node_size = 5,
  num_threads = detectCores(),
  calibrate = FALSE,
  Roy_method = "quantile",
  featureBias = TRUE,
  predictionBias = TRUE,
  Tung_R = 5,
  Tung_num_trees = 75,
  variant = 1,
  Ghosal_num_stages = 2,
  prop = 0.618,
  concise = TRUE,
  ...
)
}
\arguments{
\item{formula}{Object of class formula or character describing the model to fit. Interaction terms supported only for numerical variables.}

\item{train_data}{Training data of class data.frame, matrix, dgCMatrix (Matrix) or gwaa.data (GenABEL). Matches ranger() requirements.}

\item{test_data}{Test data of class data.frame, matrix, dgCMatrix (Matrix) or gwaa.data (GenABEL). Utilizes ranger::predict() to get prediction intervals for test data.}

\item{method}{Choose what method to generate RF prediction intervals. ("Zhang", "quantile", "Romano", "Ghosal", "Roy", "Tung", "HDI). Defaults to "Zhang".}

\item{alpha}{Significance level for prediction intervals.}

\item{symmetry}{True if constructing symmetric out-of-bag prediction intervals, False otherwise. Used only Zhang method. Defaults to TRUE.}

\item{seed}{Seed for random number generation. Currently not utilized.}

\item{m_try}{Number of variables to randomly select from at each split.}

\item{num_trees}{Number of trees used in the random forest.}

\item{min_node_size}{Minimum number of observations before split at a node.}

\item{num_threads}{The number of threads to use in parallel. Default is the current number of cores.}

\item{calibrate}{If TRUE, intervals calibrated to achieve nominal coverage. Currently uses quantiles to calibrate. Only for method = "Roy".}

\item{featureBias}{Remove feature bias. Only for method = "Tung".}

\item{predictionBias}{Remove prediction bias. Only for method = "Tung".}

\item{Tung_num_trees}{Number of trees used in bias removal. Only for method = "Tung".}

\item{variant}{Choose which variant to use. Only for method = "Ghosal". Currently variant 2 not implemented.}

\item{Ghosal_num_stages}{Number of total stages. Only for method = "Ghosal".}

\item{prop}{Proportion of training data to sample for each tree. Currently variant 2 not implemented.}

\item{concise}{If TRUE, only predictions output. Defaults to FALSE}

\item{replace}{Sample with replacement, or not. Utilized for the two different variants outlined in Ghosal, Hooker 2018. Currently variant 2 not implemented.}

\item{TungR}{Number of repetitions used in bias removal. Only for method = "Tung".}
}
\description{
Single wrapper for seven different methods, and their variants, to create random forest prediction intervals. More methods to be added.
}
\section{Values}{
NA
}

\examples{
library(piRF)

getPILength <- function(x){
#average PI length across each set of predictions
l <- x[,2] - x[,1]
avg_l <- mean(l)
return(avg_l)
}

getCoverage <- function(x, response){
  #output coverage for test data
  coverage <- sum((response >= x[,1]) * (response <= x[,2]))/length(response)
  return(coverage)
}

#import airfoil self noise dataset
data(airfoil)

#generate train and test data
ratio <- .9
nrow <- nrow(airfoil)
n <- floor(nrow*ratio)
samp <- sample(1:nrow, size = n)
train <- airfoil[samp,]
test <- airfoil[-samp,]

#generate prediction intervals
res <- rfint(pressure ~ . , train_data = train, test_data = test,
             method = c("quantile", "Zhang", "Tung", "Romano", "Roy", "HDI", "Ghosal"))

#empirical coverage, and average prediction interval length for each method
coverage <- sapply(res, FUN = getCoverage, response = test$pressure)
coverage
length <- sapply(res, FUN = getPILength)
length
}
\keyword{boosting,}
\keyword{calibration}
\keyword{forest,}
\keyword{interval,}
\keyword{prediction}
\keyword{random}
